PreLiveEventsViewController in MVVM refactor. Bugfixes. Code clean up.

+++ Core/Models/Shared/HomeFilterOptions.swift
@@ -10,16 +10,16 @@ import CoreGraphics
 
 struct HomeFilterOptions: Equatable {
 
-    let lowerBoundTimeRange: CGFloat
-    let highBoundTimeRange: CGFloat
+    let lowerBoundTimeRange: Int
+    let highBoundTimeRange: Int
     let defaultMarket: MainMarketType?
  
     let lowerBoundOddsRange: CGFloat
     let highBoundOddsRange: CGFloat
     let countFilters: Int
 
-    init(lowerBoundTimeRange: CGFloat = 0.0,
-         highBoundTimeRange: CGFloat = 365.0,
+    init(lowerBoundTimeRange: Int = 0,
+         highBoundTimeRange: Int = 365,
          defaultMarket: MainMarketType? = nil,
          lowerBoundOddsRange: CGFloat = 1.0,
          highBoundOddsRange: CGFloat = 300.0,
